Move isValidationError helper to lib
